refactor code
Removed alot of boilerplate functions together with their tests. This
boilerplate was due to parsing the season and episode signature in
movies and subtitle files without their actual value awareness (ie.
file: helloS01E04, the program was supposed to be aware that the
actual number of season and episode is 1 and 4 respectively), which
means alot of function were to be added to make it possible to compare
the actual signatures. Since now the program is aware of those values,
all of those function are not needed anymore!.